clean RecoTracker/TkNavigation
