+++ RecoEgamma/EgammaPhotonAlgos/BuildFile.xml
@@ -18,7 +18,7 @@
 <use   name="TrackingTools/PatternTools"/>
 <use   name="TrackingTools/DetLayers"/>
 <use   name="TrackingTools/TransientTrack"/>
-<use   name="RecoTracker/TkDetLayers"/>
+<use   name="RecoTracker/MeasurementDet"/>
 <use   name="RecoTracker/TkNavigation"/>
 <use   name="RecoTracker/TkTrackingRegions"/>
 <use   name="RecoTracker/CkfPattern"/>
