+++ RecoTracker/MeasurementDet/BuildFile.xml
@@ -22,4 +22,5 @@
 <use   name="RecoTracker/TransientTrackingRecHit"/>
 <use   name="TrackingTools/TrajectoryState"/>
 <use   name="TrackingTools/TransientTrackingRecHit"/>
+<use   name="RecoTracker/TkDetLayers"/>
 <use   name="rootrflx"/>
