--- RecoTracker/TkNavigation/BuildFile.xml
@@ -1,11 +1,0 @@
-<export>
-  <lib   name="1"/>
-</export>
-<use   name="TrackingTools/PatternTools"/>
-<use   name="Utilities/General"/>
-<use   name="TrackingTools/GeomPropagators"/>
-<use   name="TrackingTools/DetLayers"/>
-<use   name="DataFormats/TrackingRecHit"/>
-<use   name="DataFormats/DetId"/>
-<use   name="FWCore/PluginManager"/>
-<use   name="FWCore/ParameterSet"/>
