+++ RecoTracker/TkNavigation/plugins/NavigationSchoolESProducer.cc
@@ -1,6 +1,35 @@
 #include <FWCore/Utilities/interface/ESInputTag.h>
-#include "RecoTracker/TkNavigation/plugins/NavigationSchoolESProducer.h"
 
+#include <memory>
+#include "boost/shared_ptr.hpp"
+
+// user include files
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+
+#include "FWCore/Framework/interface/ESHandle.h"
+
+#include "NavigationSchoolFactory.h"
+#include "RecoTracker/Record/interface/NavigationSchoolRecord.h"
+
+//
+// class decleration
+//
+
+class dso_hidden NavigationSchoolESProducer final : public edm::ESProducer {
+public:
+  NavigationSchoolESProducer(const edm::ParameterSet&);
+  ~NavigationSchoolESProducer();
+  
+  typedef boost::shared_ptr<NavigationSchool> ReturnType;
+
+  virtual ReturnType produce(const NavigationSchoolRecord&);
+ protected:
+  // ----------member data ---------------------------
+  edm::ParameterSet theNavigationPSet;
+  std::string theNavigationSchoolName;
+  boost::shared_ptr<NavigationSchool> theNavigationSchool ;
+};
 
 //
 //
@@ -55,3 +84,7 @@ NavigationSchoolESProducer::produce(const NavigationSchoolRecord& iRecord)
 								    field.product()));
    return theNavigationSchool ;
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_EVENTSETUP_MODULE(NavigationSchoolESProducer);
