+++ RecoTracker/TkNavigation/plugins/SymmetricLayerFinder.h
