--- RecoTracker/TkNavigation/src/SkippingLayerCosmicNavigationSchool.cc
@@ -1,8 +1,0 @@
-#include "RecoTracker/TkNavigation/interface/SkippingLayerCosmicNavigationSchool.h"
-
-SkippingLayerCosmicNavigationSchool::SkippingLayerCosmicNavigationSchool(const GeometricSearchTracker* theInputTracker,
-									 const MagneticField* field,
-									 const CosmicNavigationSchoolConfiguration conf)
-{
-  build(theInputTracker, field, conf);
-}
